feat: remove ldev ai mcp-setup — use claude mcp add for Claude Code
The mcp-setup command generated per-project JSON config files but
duplicated what claude mcp add --scope user already does natively
(and better, since it writes to the global user config). Replace with
one-line setup instructions per editor and update docs accordingly.

For Claude Code: claude mcp add ldev --scope user -- ldev-mcp-server
For VSCode/Cursor: paste the 3-line JSON snippet from the docs

Doctor tests moved to mcp-server-doctor.test.ts with direct config
writes instead of depending on the removed setup module.
