+++ docs/agentic/index.md
@@ -54,24 +54,44 @@ AI folders rather than replacing them.
 
 ## Set up the local MCP server
 
+`ldev-mcp-server` is included in the global `@mordonezdev/ldev` install.
+
+**Claude Code** — one command, available across all projects:
+
 ```bash
-ldev ai mcp-setup --target . --tool all
+claude mcp add ldev --scope user -- ldev-mcp-server
 ```
 
-This writes the MCP config for VSCode, Claude Code and Cursor in one run.
+**VSCode** — add to `.vscode/mcp.json`:
 
-Use an explicit launch strategy when reproducibility matters:
+```json
+{
+  "servers": {
+    "ldev": { "type": "stdio", "command": "ldev-mcp-server" }
+  }
+}
+```
 
-```bash
-ldev ai mcp-setup --target . --tool vscode --strategy local
-ldev ai mcp-setup --target . --tool claude-code --strategy global
-ldev ai mcp-setup --target . --tool cursor --strategy npx
+**Cursor** — add to `.cursor/mcp.json`:
+
+```json
+{
+  "mcpServers": {
+    "ldev": { "command": "ldev-mcp-server" }
+  }
+}
+```
+
+If `ldev` is not installed globally, replace `"command": "ldev-mcp-server"` with:
+
+```json
+"command": "npx", "args": ["--package", "@mordonezdev/ldev", "-y", "ldev-mcp-server"]
 ```
 
 If the editor does not show the tools:
 
 ```bash
-ldev mcp doctor --target . --tool all
+ldev mcp doctor
 ```
 
 ## Context snapshots
