+++ src/commands/ai/ai.command.ts
@@ -3,13 +3,6 @@ import {Command} from 'commander';
 import {addOutputFormatOption, createFormattedAction} from '../../cli/command-helpers.js';
 import {formatAiResult, runAiInstall} from '../../features/ai/ai-install.js';
 import {parseBootstrapCacheTtl, runAiBootstrap} from '../../features/agent/agent-bootstrap.js';
-import {
-  formatMcpSetup,
-  runMcpSetup,
-  type McpStrategy,
-  type McpTool,
-} from '../../entrypoints/mcp-server/mcp-server-setup.js';
-
 type AiInstallCommandOptions = {
   target: string;
   force?: boolean;
@@ -18,12 +11,6 @@ type AiInstallCommandOptions = {
 type AiBootstrapCommandOptions = {
   intent: string;
   cache?: string;
-};
-
-type AiMcpSetupCommandOptions = {
-  target: string;
-  tool: McpTool;
-  strategy?: McpStrategy;
 };
 
 export function createAiCommand(): Command {
@@ -50,20 +37,6 @@ export function createAiCommand(): Command {
         'Reuse a cached bootstrap result for the same intent + cwd while it is newer than this TTL',
       ),
     'json',
-  );
-
-  const mcpSetupCommand = addOutputFormatOption(
-    command
-      .command('mcp-setup')
-      .description('Write the ldev MCP server config for your AI assistant')
-      .requiredOption('--target <target>', 'Project root to write the config into')
-      .requiredOption(
-        '--tool <tool>',
-        'AI assistant to configure: all, claude-code (.claude/mcp.json), cursor (.cursor/mcp.json), or vscode (.vscode/mcp.json)',
-      ),
-  ).option(
-    '--strategy <strategy>',
-    'Server launch strategy: global (ldev-mcp-server), local (node ./node_modules/...), or npx',
   );
 
   command.description('Standard AI assets and skills for ldev projects').addHelpText(
@@ -99,15 +72,6 @@ Meta-file bootstrap:
         cacheTtlSeconds: parseBootstrapCacheTtl(options.cache),
       });
     }),
-  );
-
-  mcpSetupCommand.action(
-    createFormattedAction(
-      async (_context, options: AiMcpSetupCommandOptions) => {
-        return runMcpSetup({targetDir: options.target, tool: options.tool, strategy: options.strategy});
-      },
-      {text: formatMcpSetup},
-    ),
   );
 
   return command;
