+++ src/entrypoints/dashboard/dashboard-worktree-routes.ts
@@ -3,7 +3,6 @@ import type http from 'node:http';
 import {loadConfig} from '../../core/config/load-config.js';
 import {formatEnvStart, runEnvStart} from '../../features/env/env-start.js';
 import {runEnvStop} from '../../features/env/env-stop.js';
-import {formatMcpSetup, runMcpSetup} from '../mcp-server/mcp-server-setup.js';
 import {formatWorktreeSetup, runWorktreeSetup} from '../../features/worktree/worktree-setup.js';
 import {readJsonBody, writeDashboardError} from './dashboard-http.js';
 import {queueDashboardTaskResponse} from './dashboard-task-commands.js';
@@ -14,7 +13,6 @@ type DashboardCreateWorktreePayload = {
   baseRef?: string;
   startAfterCreate?: boolean;
   withEnv?: boolean;
-  installMcp?: boolean;
   stopMainForClone?: boolean;
   restartMainAfterClone?: boolean;
 };
@@ -38,7 +36,6 @@ export async function handleWorktreeCreate(
     const baseRef = payload.baseRef?.trim() || undefined;
     const startAfterCreate = payload.startAfterCreate !== false;
     const withEnv = payload.withEnv !== false;
-    const installMcp = payload.installMcp !== false;
     const stopMainForClone = withEnv ? payload.stopMainForClone !== false : false;
     const restartMainAfterClone = withEnv && Boolean(payload.restartMainAfterClone);
 
@@ -59,10 +56,6 @@ export async function handleWorktreeCreate(
           printer,
         });
         printer.info(formatWorktreeSetup(result));
-        if (installMcp) {
-          const mcpResult = await runMcpSetup({targetDir: result.worktreeDir, tool: 'all'});
-          printer.info(formatMcpSetup(mcpResult));
-        }
         if (restartMainAfterClone && result.mainEnvStoppedForClone && !result.mainEnvRestartedAfterClone) {
           throw new Error(result.mainRestartError ?? 'Main environment was stopped for clone but did not restart.');
         }
